Combine WindowsRawInputHandler and WindowsMouseHandler into a simple implementation
